moved ./clmm/structures/cluster.py to ./clmm/galaxycluster.py; moved ./clmm/desc/load_from_DC2.py to ./clmm/lsst/io.py ; updated GalaxyCluster fieldnames and their corresponding use in load_from_DC2; allowed GalaxyCluster.unique_id to be string or int
