--- clmm/desc/__init__.py
@@ -1,1 +1,0 @@
-from .load_from_DC2 import *
