--- clmm/structures/__init__.py
@@ -1,1 +1,0 @@
-from .cluster import *
